+++ redash/query_runner/__init__.py
@@ -10,10 +10,12 @@ from urllib.parse import urlparse
 from six import text_type
 from sshtunnel import open_tunnel
 from redash import settings, utils
-from redash.utils import json_loads, query_is_select_no_limit, add_limit_to_query
+from redash.utils import json_loads
 from rq.timeouts import JobTimeoutException
 
 from redash.utils.requests_session import requests, requests_session
+
+import sqlparse
 
 logger = logging.getLogger(__name__)
 
@@ -48,6 +50,61 @@ SUPPORTED_COLUMN_TYPES = set(
     [TYPE_INTEGER, TYPE_FLOAT, TYPE_BOOLEAN, TYPE_STRING, TYPE_DATETIME, TYPE_DATE]
 )
 
+def split_sql_statements(query):
+    def strip_trailing_comments(stmt):
+        idx = len(stmt.tokens) - 1
+        while idx >= 0:
+            tok = stmt.tokens[idx]
+            if tok.is_whitespace or sqlparse.utils.imt(tok, i=sqlparse.sql.Comment, t=sqlparse.tokens.Comment):
+                stmt.tokens[idx] = sqlparse.sql.Token(sqlparse.tokens.Whitespace, ' ')
+            else:
+                break
+            idx -= 1
+        return stmt
+
+    def strip_trailing_semicolon(stmt):
+        idx = len(stmt.tokens) - 1
+        while idx >= 0:
+            tok = stmt.tokens[idx]
+            # we expect that trailing comments already are removed
+            if not tok.is_whitespace:
+                if sqlparse.utils.imt(tok, t=sqlparse.tokens.Punctuation) and tok.value == ";":
+                    stmt.tokens[idx] = sqlparse.sql.Token(sqlparse.tokens.Whitespace, ' ')
+                break
+            idx -= 1
+        return stmt
+
+    def is_empty_statement(stmt):
+        strip_comments = sqlparse.filters.StripCommentsFilter()
+
+        # copy statement object. `copy.deepcopy` fails to do this, so just re-parse it
+        st = sqlparse.engine.FilterStack()
+        stmt = next(st.run(sqlparse.text_type(stmt)))
+
+        sql = sqlparse.text_type(strip_comments.process(stmt))
+        return sql.strip() == ""
+
+    stack = sqlparse.engine.FilterStack()
+
+    result = [stmt for stmt in stack.run(query)]
+    result = [strip_trailing_comments(stmt) for stmt in result]
+    result = [strip_trailing_semicolon(stmt) for stmt in result]
+    result = [sqlparse.text_type(stmt).strip() for stmt in result if not is_empty_statement(stmt)]
+
+    if len(result) > 0:
+        return result
+
+    return [""]  # if all statements were empty - return a single empty statement
+
+
+def combine_sql_statements(queries):
+    return ";\n".join(queries)
+
+def find_last_keyword_idx(parsed_query):
+    for i in reversed(range(len(parsed_query.tokens))):
+        if parsed_query.tokens[i].ttype in sqlparse.tokens.Keyword:
+            return i
+    return -1
 
 class InterruptException(Exception):
     pass
@@ -223,14 +280,40 @@ class BaseSQLQueryRunner(BaseQueryRunner):
     def supports_auto_limit(self):
         return True
 
+    @property
+    def limit_text(self):
+        return " LIMIT 1000"
+
+    def query_is_select_no_limit(self, query):
+        parsed_query = sqlparse.parse(query)[0]
+        last_keyword_idx = find_last_keyword_idx(parsed_query)
+        # Either invalid query or query that is not select
+        if last_keyword_idx == -1 or parsed_query.tokens[0].value.upper() != "SELECT":
+            return False
+
+        no_limit = parsed_query.tokens[last_keyword_idx].value.upper() != "LIMIT" \
+                and parsed_query.tokens[last_keyword_idx].value.upper() != "OFFSET"
+
+        return no_limit
+
+    def add_limit_to_query(self, query):
+        parsed_query = sqlparse.parse(query)[0]
+        limit_tokens = sqlparse.parse(self.limit_text)[0].tokens
+        length = len(parsed_query.tokens)
+        if parsed_query.tokens[length - 1].ttype == sqlparse.tokens.Punctuation:
+            parsed_query.tokens[length - 1:length - 1] = limit_tokens
+        else:
+            parsed_query.tokens += limit_tokens
+        return str(parsed_query)
+
+
     def apply_auto_limit(self, query_text, should_apply_auto_limit):
         if should_apply_auto_limit:
-            from redash.query_runner.databricks import split_sql_statements, combine_sql_statements
             queries = split_sql_statements(query_text)
             # we only check for last one in the list because it is the one that we show result
             last_query = queries[-1]
-            if query_is_select_no_limit(last_query, self.type()):
-                queries[-1] = add_limit_to_query(last_query, self.type())
+            if self.query_is_select_no_limit(last_query):
+                queries[-1] = self.add_limit_to_query(last_query)
             return combine_sql_statements(queries)
         else:
             return query_text
